Fixed mocking requests.post
